AI Project 2
In this task, you will implement a branch-and-bound search for finding
the best set of nodes to seed.  In your myagent.py, selectNodes()
method, you will see a skeleton which you are to complete.  The skeleton
suggests an implementation of branchand-bound
search using the familiar outlines of the search algorithms that you
implemented in project 1 and described in the textbook (DFS, A*, etc).
This implementation is a bit simpler, although it is perhaps more common
to implement
it as a recursive tree search.